Testing: Combines queue/compute that have effectively duplicates names
